steps a list

+++ include/render.h
@@ -2,8 +2,8 @@
 #define RENDER
 extern bool sprites[128];
 extern bool spritesSub[128];
-u8 popSprite();
-u8 popSpriteSub();
-void pushSprite(u8 i);
-void pushSpriteSub(u8 i);
+int popSprite();
+int popSpriteSub();
+void pushSprite(int i);
+void pushSpriteSub(int i);
 #endif
